pyproject.toml
[tool.poetry]
name = "longliving"
version = "0.1.0"
description = ""
authors = ["Igor Davidyuk <igor.davidyuk@intel.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.20.1"
grpcio-tools = "^1.36.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
